Tried some fixes for USI Rosa, but proper flags not found
The code is very slow on Rosa. I believe it's because I do not have the right flags for the Infiniband connection. But no matter what I tried, no IB device seemed to be found. So, the code is incredibly slow to execute...about 6x slower than on an M1 MacBook Pro.